pip_requirements.txt
numpy==1.20.3
jsonlines==2.0.0
torchmetrics==0.7.3
pytorch-lightning==1.4.9
pandas==1.3.2
pywikibot==6.5.0
wikitextparser==0.47.5
wandb==0.12.2
omegaconf==2.2.2
hydra-core==1.2.0
hydra-colorlog==1.1.0
rich==10.10.0
pre-commit==2.15.0
matplotlib==3.4.2
tqdm==4.62.2
jupyter==1.0.0
ipykernel==6.4.1
jupyterlab==3.1.7
einops==0.4.1
scikit-learn==1.1.0
pyyaml==6.0
runai==0.4.1
datasets==2.2.2
detoxify==0.5.0
pytest-pycharm==0.7.0
pytest-subtests==0.8.0
seaborn==0.11.2
sacrebleu==2.2.0
chex==0.1.3
networkx==2.8
scipy==1.8.1
jax==0.3.15
jaxlib==0.3.15
sentencepiece==0.1.96
transformers==4.10.0
protobuf==3.20
setuptools==59.5.0
